+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 <unreleased>
 
+- Multiple input files can be given on the command line
+- Argument `--outdir` sets the output directory
+- Argument `--project` defines TSCOnfig for the source files
 - Accept instantiation expressions `func<TYPE>`
 
 Version 0.7.0 - 2025-02-02